chore: extract associations and organize imports
