+++ package.json
@@ -11,7 +11,7 @@
   "scripts": {
     "babel-node": "babel-node --presets='@babel/preset-env'",
     "start": "nodemon --exec npm run babel-node -- ./src/server.js",
-    "seed": "node ./scripts/seedDb.js",
+    "seed": "npm run babel-node -- ./scripts/seedDb.js",
     "build": "babel src -d lib"
   },
   "devDependencies": {
